+++ airflow-core/src/airflow/api/common/mark_tasks.py
@@ -25,8 +25,8 @@ from typing import TYPE_CHECKING, TypeAlias
 from sqlalchemy import and_, or_, select
 from sqlalchemy.orm import lazyload
 
-from airflow.models.dagrun import DagRun
-from airflow.models.taskinstance import DagRunState, TaskInstance, TaskInstanceState
+from airflow.models.dagrun import DagRun, DagRunState
+from airflow.models.taskinstance import TaskInstance, TaskInstanceState
 from airflow.utils.session import NEW_SESSION, provide_session
 from airflow.utils.state import State
