+++ airflow-core/src/airflow/models/dagrun.py
@@ -23,13 +23,13 @@ import re
 from collections import defaultdict
 from collections.abc import Callable, Iterable, Iterator, Sequence
 from datetime import datetime
+from enum import Enum
 from typing import TYPE_CHECKING, Any, NamedTuple, TypeVar, cast, overload
 
 import structlog
 from natsort import natsorted
 from sqlalchemy import (
     JSON,
-    Enum,
     ForeignKey,
     ForeignKeyConstraint,
     Index,
@@ -64,7 +64,7 @@ from airflow.listeners.listener import get_listener_manager
 from airflow.models import Deadline, Log
 from airflow.models.backfill import Backfill
 from airflow.models.base import Base, StringID
-from airflow.models.taskinstance import DagRunState, TaskInstance as TI, TaskInstanceState
+from airflow.models.taskinstance import TaskInstance as TI, TaskInstanceState
 from airflow.models.taskinstancehistory import TaskInstanceHistory as TIH
 from airflow.models.tasklog import LogTemplate
 from airflow.models.taskmap import TaskMap
@@ -116,6 +116,24 @@ if TYPE_CHECKING:
 RUN_ID_REGEX = r"^(?:manual|scheduled|asset_triggered)__(?:\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\+00:00)$"
 
 log = structlog.get_logger(__name__)
+
+
+class DagRunState(str, Enum):
+    """
+    All-possible states that a DagRun can be in.
+
+    These are "shared" with TaskInstanceState in some parts of the code,
+    so please ensure that their values always match the ones with the
+    same name in TaskInstanceState.
+    """
+
+    QUEUED = "queued"
+    RUNNING = "running"
+    SUCCESS = "success"
+    FAILED = "failed"
+
+    def __str__(self) -> str:
+        return self.value
 
 
 class TISchedulingDecision(NamedTuple):
