+++ airflow-core/tests/unit/api/common/test_mark_tasks.py
@@ -22,8 +22,8 @@ import pytest
 from sqlalchemy import select
 
 from airflow.api.common.mark_tasks import set_dag_run_state_to_failed, set_dag_run_state_to_success
-from airflow.models.dagrun import DagRun
-from airflow.models.taskinstance import DagRunState, TaskInstanceState
+from airflow.models.dagrun import DagRun, DagRunState
+from airflow.models.taskinstance import TaskInstanceState
 from airflow.providers.standard.operators.empty import EmptyOperator
 from airflow.utils.state import State
