+++ airflow-core/tests/unit/api_fastapi/core_api/routes/public/test_task_instances.py
@@ -36,9 +36,10 @@ from airflow.jobs.triggerer_job_runner import TriggererJobRunner
 from airflow.listeners.listener import get_listener_manager
 from airflow.models import DagRun, Log, TaskInstance
 from airflow.models.dag_version import DagVersion
+from airflow.models.dagrun import DagRunState
 from airflow.models.hitl import HITLDetail
 from airflow.models.renderedtifields import RenderedTaskInstanceFields as RTIF
-from airflow.models.taskinstance import DagRunState, TaskInstanceState
+from airflow.models.taskinstance import TaskInstanceState
 from airflow.models.taskinstancehistory import TaskInstanceHistory
 from airflow.models.taskmap import TaskMap
 from airflow.models.trigger import Trigger
