+++ airflow-core/tests/unit/api_fastapi/core_api/routes/ui/test_dashboard.py
@@ -24,7 +24,8 @@ import pytest
 
 from airflow.models.dag import DagModel
 from airflow.models.dagbag import DBDagBag
-from airflow.models.taskinstance import DagRunState, TaskInstanceState
+from airflow.models.dagrun import DagRunState
+from airflow.models.taskinstance import TaskInstanceState
 from airflow.providers.standard.operators.empty import EmptyOperator
 from airflow.utils.types import DagRunType
