+++ airflow-core/tests/unit/listeners/test_listeners.py
@@ -26,7 +26,8 @@ from airflow._shared.timezones import timezone
 from airflow.exceptions import AirflowException
 from airflow.jobs.job import Job, run_job
 from airflow.listeners.listener import get_listener_manager
-from airflow.models.taskinstance import DagRunState, TaskInstanceState
+from airflow.models.dagrun import DagRunState
+from airflow.models.taskinstance import TaskInstanceState
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.utils.session import provide_session
