+++ airflow-core/tests/unit/models/test_backfill.py
@@ -37,7 +37,8 @@ from airflow.models.backfill import (
     ReprocessBehavior,
     _create_backfill,
 )
-from airflow.models.taskinstance import DagRunState, TaskInstanceState
+from airflow.models.dagrun import DagRunState
+from airflow.models.taskinstance import TaskInstanceState
 from airflow.providers.standard.operators.python import PythonOperator
 from airflow.ti_deps.dep_context import DepContext
 from airflow.utils.types import DagRunTriggeredByType, DagRunType
