+++ airflow-core/tests/unit/ti_deps/deps/test_prev_dagrun_dep.py
@@ -23,7 +23,8 @@ from unittest.mock import ANY, Mock, patch
 import pytest
 
 from airflow._shared.timezones.timezone import convert_to_utc, datetime
-from airflow.models.taskinstance import DagRunState, TaskInstanceState
+from airflow.models.dagrun import DagRunState
+from airflow.models.taskinstance import TaskInstanceState
 from airflow.sdk import DAG, BaseOperator
 from airflow.ti_deps.dep_context import DepContext
 from airflow.ti_deps.deps.prev_dagrun_dep import PrevDagrunDep
